Refactor the component tree inside the MainOrganiser component namig to make more sense
